Update Functionality
So, this merge is pretty buggy, but need to push it so we have something to show for tonight's assignment.

BUG:
For some reason, when the router submits a put request, it also submits a POST request. At best, it updates an entity and adds the same one to the database (so it appears twice). At worst it crashes because the POST request doesn't have the proper information it needs to create a new instance.